chore: cargo clipply

+++ scales/src/value.rs
@@ -260,14 +260,12 @@ fn sequence_size(data: &[u8]) -> (usize, usize) {
         match len {
             1 => (data[0] >> 2).into(),
             2 => u16::from(data[0] >> 2 | data[1] << 6).into(),
-            4 => u32::from(
-                (data[0] as u32) >> 2
-                    | (data[1] as u32) << 6
-                    | (data[2] as u32) << 14
-                    | (data[3] as u32) << 22,
-            )
-            .try_into()
-            .unwrap(),
+            4 => ((data[0] as u32) >> 2
+                | (data[1] as u32) << 6
+                | (data[2] as u32) << 14
+                | (data[3] as u32) << 22)
+                .try_into()
+                .unwrap(),
             _ => todo!(),
         },
         len,
